Harden research publication boundaries
